Correct the behavior of DrawJoystickDot in SDL 2.0
In SDL 1.2, the minimum value that a joystick axis can be at is -32767.
In SDL 2, the minimum value is actually the minimum value of type
Sint16, which is -32768.

Previously the joystick value had 32767 added to it to form the screen
coordinate of the dot, which was cast into (Uint32) -1, or +4 billion.